@@ -1,18 +1,20 @@
-'use strict';
-const path = require('path');
-const assert = require('yeoman-assert');
-const helpers = require('yeoman-test');
+/* eslint-disable jest/expect-expect */
+import path from 'path';
+import assert from 'yeoman-assert';
+import helpers from 'yeoman-test';
 
 const name = 'test-project';
+const email = 'alex@test.com';
 const description = 'description of test project';
 const username = 'alanturing';
 const fullname = 'Alan Turing';
+const license = 'MIT';
 
-describe('generator-npm-typescript-starter:app', () => {
+describe('generator-typescript-library-starter:app', () => {
   beforeAll(() => {
     return helpers
-      .run(path.join(__dirname, '../generators/app'))
-      .withPrompts({ name, description, username, fullname });
+      .run(path.join(__dirname, '../app'))
+      .withPrompts({ name, description, username, email, fullname, license });
   });
 
   it('creates files', () => {
@@ -21,14 +23,11 @@ describe('generator-npm-typescript-starter:app', () => {
       'package.json',
       'README.md',
       '.gitignore',
-      '.nycrc',
       'index.ts',
-      'mocha.opts',
       'tsconfig.json',
-      'tslint.json',
-      'lib/index.ts',
-      'test/index.spec.ts',
-      '.vscode/settings.json'
+      '.eslintrc.js',
+      'src/index.ts',
+      'src/index.spec.ts',
     ]);
   });
 
